Refactor MapMagic brush margins to use world space
- Changed `margins` from `int` to `float` in `Preset.cs` and `Stamp.cs` to make it world space instead of pixel-based.
- Updated `Min`, `Max`, and `Size` properties in `Stamp` to correctly include the `margins` value.
- Converted `margins` to pixel integer during `Area` construction in `GetArea` using `Mathf.CeilToInt(margins / pixelSize.x)`.
- Removed related TODO comment as it is now resolved.

+++ Assets/MapMagic/Brush/Core/Preset.cs
@@ -25,7 +25,7 @@ namespace MapMagic.Brush
 
 		public float radius = 100;
 		public float hardness = 0.5f;
-		public int margins = 0;
+		public float margins = 0;
 		public float spacing = 0.25f;
 
 		public Override ovd = new Override();
